requirements.txt
celery
channels
channels-redis
crispy-bootstrap5
daphne
Django
django-celery-beat
django-celery-results
django-ckeditor
django-countries
django-crispy-forms
django-dbbackup
django-extensions
django-filter
django-forestadmin
django-furl
django-import-export
django-ipware
django-jazzmin
django-logentry-admin
django-material-admin
django-mptt
django-parler
django-pghistory
django-pgtrigger
django-phonenumber-field
django-rest-swagger
django-rosetta
django-tabular-permissions
django-widget-tweaks
djangochannelsrestframework
djangorestframework
djangorestframework-recursive
djangorestframework-simplejwt
djangorestframework-word-filter
djoser
drf-dynamic-fields
drf-extensions
drf-spectacular
drf-spectacular-sidecar
drf-tracking
drf-yasg
easy-thumbnails
gunicorn
import-export
numpy
pandas
phonenumbers
Pillow
psycopg2-binary
redis
uvicorn
django-environ
sentry-sdk
furl
drf-api-logger
django-auditlog
django-haystack[elasticsearch]
djangorestframework-csv
whoosh